refactor: unify SQL generation and agent logic by replacing KowalskiSQLAgent with an enhanced KowalskiAgent implementation
